feat : 지각 시작, 출석 시작
