+++ apps/admin/components/attendance/ExtraScoreModal.tsx
@@ -1,5 +1,5 @@
 import { useAttendanceAdditionalScore } from '@weekly/api';
-import { AttendanceSessionResponse } from '@weekly/api/lib/types/attendance';
+import { AttendanceSessionResponse } from '@weekly/api/lib/types/admin';
 import { Button, Input, styled } from '@weekly/ui';
 import { useValidateState } from '@weekly/utils';
 import { useRouter } from 'next/router';
