+++ packages/api/lib/admin/queries/index.ts
@@ -1,5 +1,6 @@
 import useAdminLogin from './useAdminLogin';
 import useAttendanceAdditionalScore from './useAttendanceAdditionalScore';
+import useAttendanceQr from './useAttendanceQr';
 import useAttendanceSession from './useAttendanceSession';
 import useCreateGeneration from './useCreateGeneration';
 import useDeleteGeneration from './useDeleteGeneration';
@@ -10,6 +11,7 @@ import useSessionByGeneration from './useSessionByGeneration';
 export {
   useAdminLogin,
   useAttendanceAdditionalScore,
+  useAttendanceQr,
   useAttendanceSession,
   useCreateGeneration,
   useDeleteGeneration,
