+++ packages/api/lib/types/attendance.ts
@@ -95,26 +95,3 @@ export interface FindAttendanceProfileResponse {
   isGenerationMember: boolean;
   attendanceData?: AttendanceProfileResponse;
 }
-
-export interface AttendanceSessionResponses {
-  week: number;
-  sessionDate: string;
-  attended: number;
-  tardy: number;
-  absence: number;
-  data: AttendanceSessionResponse[];
-}
-
-export interface AttendanceSessionResponse {
-  name: string;
-  number: string;
-  attendanceId: number;
-  position?: string;
-  subPosition?: string;
-  initialGeneration: number;
-  scoreChanged: number;
-  score?: number;
-  attendanceStatus: string;
-  extraScoreNote?: string;
-  note?: string;
-}
